+++ example/lib/src/domain/model/deezer_track.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'deezer_track.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_DeezerTrack _$$_DeezerTrackFromJson(Map<String, dynamic> json) =>
+    _$_DeezerTrack(
+      id: json['id'] as int,
+      title: json['title'] as String,
+      url: json['preview'] as String,
+    );
+
+Map<String, dynamic> _$$_DeezerTrackToJson(_$_DeezerTrack instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'title': instance.title,
+      'preview': instance.url,
+    };